requirements.txt
viam-sdk>=0.4.0
Pillow>=9.0.0
adafruit-circuitpython-rgb-display>=3.0.0
